Replaces Int values by Long (love "id")

+++ app/src/main/java/ovh/geoffrey_druelle/typicodephotos/data/remote/model/Photo.kt
@@ -3,8 +3,8 @@ package ovh.geoffrey_druelle.typicodephotos.data.remote.model
 import com.google.gson.annotations.SerializedName
 
 data class Photo(
-    @SerializedName("albumId") val albumId: Int,
-    @SerializedName("id") val id: Int,
+    @SerializedName("albumId") val albumId: Long,
+    @SerializedName("id") val id: Long,
     @SerializedName("thumbnailUrl") val thumbnailUrl: String,
     @SerializedName("title") val title: String,
     @SerializedName("url") val url: String
